requirements.txt
django==1.11.2
celery==4.0.2
google-cloud==0.25.0
